+++ components/internal/StacFields.vue
@@ -1,7 +1,7 @@
 <template>
 	<section class="vue-component stac stac-fields metadata">
 		<template v-for="group in fields">
-			<component :is="headingTag" v-html="group.label" :key="group.extension" />
+			<component :is="headingTag" v-html="group.label || 'General'" :key="group.extension" />
 			<div v-for="(prop, field) in group.properties" :key="group.extension + field" :id="'field_' + field" class="tabular" :class="{wrap: Boolean(prop.custom || prop.items)}">
 				<label :title="field" v-html="prop.label" />
 				<div class="value">